requirements.txt
Django==3.2.0
python-decouple==3.6
django-email-verification==0.1.0
django-crispy-forms==1.14.0
django-localflavor==3.1
django-model-utils==4.2.0
django-rest-swagger==2.2.0
django-seed==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
pytest==7.0.1
pytest-django==4.5.2
requests==2.27.1
drf-yasg==1.20.0
psycopg2-binary==2.9.3
openapi-codec==1.3.2
django-anymail==8.5
django-seed==0.3.1
django-model-utils==4.2.0
django-cors-headers==3.11.0
django-allauth==0.50.0
django-rest-auth==0.9.5
mysqlclient==2.1.0